wancheng4

+++ src/main/java/com/zqw/order/manage/model/Spread.java
@@ -2,25 +2,25 @@ package com.zqw.order.manage.model;
 
 import com.zqw.order.manage.domain.p.BasePage;
 
+import java.math.BigInteger;
+
 /**
  * @author: zhuquanwen
  * @date: 2017/11/14 9:17
  * @desc:
  */
 public class Spread extends BasePage {
-    private Long goodsId;
+    private BigInteger goodsId;
     private String goodsName;
     private String url;
     private String userName;
 
-    public Long getGoodsId() {
+    public BigInteger getGoodsId() {
         return goodsId;
     }
-
-    public void setGoodsId(Long goodsId) {
+    public void setGoodsId(BigInteger goodsId) {
         this.goodsId = goodsId;
     }
-
     public String getGoodsName() {
         return goodsName;
     }
